lite refactor for util functions in observequery
