+++ packages/api-graphql/src/utils/index.ts
@@ -3,3 +3,5 @@
 
 export { resolveConfig } from './resolveConfig';
 export { resolveLibraryOptions } from './resolveLibraryOptions';
+export { resolvePKFields } from './resolvePKFields';
+export { findIndexByFields } from './findIndexByFields';
